fix handling equipable items to work with duplicate characters
Previously encountered bug that when duplicate characters is enabled,
if a frog duplicate equips Hero Medal or Grand Leon, and is not the "Frog"
character (e.g. a "Marle" who is a frog duplicate), then tracker
will not show the item anymore. This fixes that by scanning all
characters for the item, to work with duplicate characters flag.