create toggle mechanism for showing extra items for PopTracker
Add "Show Extra Options" in settings menu for PopTracker to
dynamically change items grid to add or remove extra items. Dynamic
layouts and the ScriptHost:AddWatchForCode mechanisms are not
supported in EmoTracker, so only enable this for PopTracker.
Disable "options" view for EmoTracker for now, as it supports
none for this pack atm.